pyproject.toml
[project]
name = "shvcli"
version = "0.6.2"
description = "Silicon Heaven CLI access application"
readme = "README.rst"
license = {file = "LICENSE"}
authors = [
  { name="Elektroline a.s." },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Topic :: Home Automation",
  "Topic :: System :: Shells",
  "Topic :: Terminals",
]
requires-python = ">=3.11"
dependencies = [
  "pyshv >= 0.8.0",
  "prompt_toolkit",
  "pygments",
  "pyxdg",
  "more-itertools",
]

[project.urls]
repository = "https://gitlab.com/silicon-heaven/shvcli"

[project.scripts]
shvcli = "shvcli.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.coverage.run]
branch = true
[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if typing.TYPE_CHECKING:"
]

[tool.ruff]
target-version = "py311"
[tool.ruff.lint]
select = [
  "F", "E", "W", "I", "N", "D", "UP", "YTT", "ANN", "ASYNC", "ASYNC1", "S", "BLE",
  "B", "A", "COM", "PL", "RUF"
]
ignore = [
  # Magic methods have defined behavior so what is there to document. __init__
  # should be documented in class docstring.
  "D105", "D107",
  # These have conflicting rules.
  "D203", "D213",
  # Ignore too long lines (format tackles that)
  "E501",
  # It is nice that assert should not be used for input checking but it can be
  # also used for internal consistency checking.
  "S101", "B011",
  # subprocess can be insecurely used but not is insecure by design
  "S404", "S603", "S607",
  # Catching Exception is valid and safe in some cases and thus should be allowed
  "BLE001",
  # Conflict with formatter
  "COM812",
  # Remove too-many limit for more complex code.
  "PLR0904", "PLR0911", "PLR0912", "PLR0913", "PLR0917", "PLR2004", "PLR1702",
  "PLR0916",
  # This seems like a wrong rule
  "PLW1641",
]
preview = true
[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "D100", "D101", "D102", "D103", "D104", "S101", "S106", "ANN001", "ANN002",
  "ANN003", "ANN201", "ANN202", "ANN204", "ANN205", "PLR6301",
]
"docs/conf.py" = ["D100", "D103", "A001", "ANN201", "ANN001"]
[tool.ruff.format]
docstring-code-format = true
preview = true

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
